Apply the suggestions (#8)
Leave dose_uncertainty.py as it is impossible to check, has not been changed/modularized yet. It will be taken care of after completing the NN generation part.
